Finished control bar logic and added harmony drag sources

+++ Sequencer/Timer.cpp
@@ -69,10 +69,10 @@ namespace System
             *status = true;
     }
 
-    void Timer::setTempo(double microseconds)
+    void Timer::setTempo(int microseconds)
     {
         double divisionsPerBeat = static_cast<double>(Music::Sixteenth) / static_cast<double>(application->song->timeDivision);
-        double microsecondsPerTick = microseconds / divisionsPerBeat / application->ticksPerDivision;
+        double microsecondsPerTick = static_cast<double>(microseconds) / divisionsPerBeat / application->ticksPerDivision;
         tickLength = std::chrono::nanoseconds(static_cast<int>(round(microsecondsPerTick * 1000)));
     }
 }
